# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GoKarabakh API",
    "description": "Booking platform for tours, hotels, and guides in Karabakh",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:5000" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "creation_date": {"type": "string", "format": "date-time"},
          "modification_date": {"type": "string", "format": "date-time"},
          "kx_count": {"type": "number"},
          "name": {"type": "string"},
          "year_of_birth": {"type": "integer"},
          "month_of_birth": {"type": "integer"},
          "day_of_birth": {"type": "integer"},
          "email": {"type": "string", "format": "email"},
          "role": {"type": "string", "enum": ["user", "owner", "guide"]}
        }
      },
      "Booking": {
        "allOf": [
          {"$ref": "#/components/schemas/Place"},
          {"type": "object", "properties": {"booked_at": {"type": "string", "format": "date-time"}}}
        ]
      },
...
      "KxTransaction": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "creation_date": {"type": "string", "format": "date-time"},
          "modification_date": {"type": "string", "format": "date-time"},
          "user_id": {"type": "string"},
          "amount": {"type": "number", "description": "Positive = earned, negative = spent"},
          "reason": {"type": "string"}
        }
      },
...
      "Error": {
        "type": "object",
        "properties": { "error": {"type": "string"} }
      }
    }
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in and get a JWT access token",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["email", "password"],
            "properties": {
              "email": {"type": "string", "format": "email"},
              "password": {"type": "string", "format": "password"}
            }
          }}}
        },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": {
            "type": "object", "properties": {"access_token": {"type": "string"}, "user": {"$ref": "#/components/schemas/User"}}
          }}}},
          "401": { "description": "Invalid email or password", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "tags": ["Users"], "summary": "Get the logged-in user's own profile",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Current user", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/User"}}}}
        }
      }
    },
    "/api/v1/users/kx_transactions": {
      "get": {
        "tags": ["Users"], "summary": "List the logged-in user's kx (loyalty points) transaction history, newest first",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Transactions", "content": { "application/json": { "schema": {"type": "array", "items": {"$ref": "#/components/schemas/KxTransaction"}}}}}
        }
      }
    },
    "/api/v1/community/traces": {
      "post": {
        "tags": ["Community Traces"], "summary": "Leave a trace (short note) about a region",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["region_slug", "caption"],
            "properties": {
              "region_slug": {"type": "string", "enum": ["shusha", "kalbajar", "lachin", "khankendi", "aghdam", "khojaly", "khojavend", "qubadli", "zangilan"]},
              "caption": {"type": "string"},
              "photo_url": {"type": "string"}
            }
          }}}
        },
        "responses": {
          "200": { "description": "Trace created", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/CommunityTrace"}}}},
          "400": { "description": "Missing/invalid fields", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      },
      "get": {
        "tags": ["Community Traces"], "summary": "List community traces, newest first",
        "parameters": [{"name": "region", "in": "query", "required": false, "schema": {"type": "string"}, "description": "Filter by region_slug"}],
        "responses": {
          "200": { "description": "Traces", "content": { "application/json": { "schema": {"type": "array", "items": {"$ref": "#/components/schemas/CommunityTrace"}}}}}
        }
      }
    },
    "/api/v1/community/traces/{id}": {
      "delete": {
        "tags": ["Community Traces"], "summary": "Remove a trace (its author or an admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": { "description": "Removed" },
          "403": { "description": "Not the author or an admin", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Error"}}}},
          "404": { "description": "Trace not found" }
        }
      }
    },
    "/api/v1/users/create": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["email", "password", "about_me"],
            "properties": {
              "email": {"type": "string", "format": "email"},
              "password": {"type": "string", "format": "password", "minLength": 8},
              "about_me": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {"type": "string"},
                  "year_of_birth": {"type": "integer"},
                  "month_of_birth": {"type": "integer"},
                  "day_of_birth": {"type": "integer"}
                }
              }
            }
          },
          "example": {
            "email": "user@example.com", "password": "hunter22",
            "about_me": {"name": "Tamerlan", "year_of_birth": 2005, "month_of_birth": 6, "day_of_birth": 14}
          }}}
        },
        "responses": {
          "200": { "description": "User created with role=user. Use /api/v1/partner-applications to apply for owner/guide access." },
          "400": { "description": "Missing/invalid fields, or password under 8 characters", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Error"}}}},
          "409": { "description": "Email is already registered", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
...
    "/api/v1/users/get_id": {
      "get": {
        "tags": ["Users"], "summary": "Get a user by id",
        "parameters": [{"name": "id", "in": "query", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": { "description": "User found", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/User"}}}},
          "400": { "description": "Missing id" },
          "404": { "description": "No user with specified id" }
        }
      }
    },
    "/api/v1/users/update": {
      "put": {
        "tags": ["Users"], "summary": "Update the logged-in user's profile",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object",
            "properties": {
              "name": {"type": "string"},
              "year_of_birth": {"type": "integer"},
              "month_of_birth": {"type": "integer"},
              "day_of_birth": {"type": "integer"}
            }
          }}}
        },
        "responses": {
          "200": { "description": "Updated user", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/User"}}}},
          "400": { "description": "No valid fields to update" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/v1/users/delete": {
      "post": {
        "tags": ["Users"], "summary": "Delete the logged-in user's account",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["password"],
            "properties": {"password": {"type": "string", "format": "password"}}
          }}}
        },
        "responses": {
          "200": { "description": "User deleted" },
          "401": { "description": "Invalid password" }
        }
      }
    },
...
    "/api/v1/places/get_id": {
      "get": {
        "tags": ["Places"], "summary": "Get a place by id",
        "parameters": [{"name": "id", "in": "query", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": { "description": "Place found", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Place"}}}},
          "404": { "description": "No place with specified id" }
        }
      }
    },
    "/api/v1/places/get_places_by_tags": {
      "get": {
        "tags": ["Places"], "summary": "Get places matching all given tags",
        "parameters": [{"name": "tags", "in": "query", "required": true, "schema": {"type": "array", "items": {"type": "string"}}, "style": "form", "explode": true}],
        "responses": {
          "200": { "description": "Matching places", "content": { "application/json": { "schema": {"type": "array", "items": {"$ref": "#/components/schemas/Place"}}}}},
          "400": { "description": "Tags were not provided" }
        }
      }
    },
    "/api/v1/places/create": {
      "post": {
        "tags": ["Places"], "summary": "Create a new place (tour, hotel, or guide listing). Requires a owner or guide account.",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["name", "description", "cost", "main_photo_url"],
            "properties": {
              "name": {"type": "string"},
              "description": {"type": "string"},
              "cost": {"type": "number"},
              "is_tour": {"type": "boolean"},
              "main_photo_url": {"type": "string"},
              "tags": {"type": "array", "items": {"type": "string"}}
            }
          },
          "example": {
            "name": "Shusha old town walking tour",
            "description": "Half-day guided walk through Shusha's historic quarter",
            "cost": 25, "is_tour": true,
            "main_photo_url": "https://example.com/photo.jpg",
            "tags": ["tour", "history"]
          }}}
        },
        "responses": {
          "200": { "description": "Place created", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Place"}}}},
          "400": { "description": "Missing required fields" },
          "403": { "description": "Account is not a owner or guide", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
    "/api/v1/places/update": {
      "put": {
        "tags": ["Places"], "summary": "Update a place you own",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["id"],
            "properties": {
              "id": {"type": "string"},
              "name": {"type": "string"},
              "description": {"type": "string"},
              "cost": {"type": "number"},
              "is_tour": {"type": "boolean"},
              "main_photo_url": {"type": "string"},
              "tags": {"type": "array", "items": {"type": "string"}}
            }
          }}}
        },
        "responses": {
          "200": { "description": "Updated place", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Place"}}}},
          "400": { "description": "Missing id or no valid fields" },
          "403": { "description": "Not the owner" },
          "404": { "description": "No place with specified id" }
        }
      }
    },
    "/api/v1/places/delete": {
      "post": {
        "tags": ["Places"], "summary": "Delete a place you own",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {"type": "object", "required": ["id"], "properties": {"id": {"type": "string"}}}}}
        },
        "responses": {
          "200": { "description": "Place deleted" },
          "403": { "description": "Not the owner" },
          "404": { "description": "No place with specified id" }
        }
      }
    },
    "/api/v1/places/book": {
      "post": {
        "tags": ["Places"], "summary": "Book a place",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["place_id"],
            "properties": { "place_id": {"type": "string"}, "use_kx": {"type": "boolean"} }
          }}}
        },
        "responses": {
          "200": { "description": "Booking confirmed", "content": { "application/json": { "schema": {
            "type": "object", "properties": {"status": {"type": "string"}, "amount_charged": {"type": "number"}}
          }}}},
          "404": { "description": "User or place not found" }
        }
      }
    },
    "/api/v1/users/bookings": {
      "get": {
        "tags": ["Users"], "summary": "List the logged-in user's booking history",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Booked places with booking timestamps", "content": { "application/json": { "schema": {"type": "array", "items": {"$ref": "#/components/schemas/Booking"}}}}}
        }
      }
    },
    "/api/v1/places/stats": {
      "get": {
        "tags": ["Places"], "summary": "Booking counts and revenue for the logged-in owner's/guide's places",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Per-place and total stats", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/PlaceStats"}}}},
          "403": { "description": "Account is not a owner or guide", "content": { "application/json": { "schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
...
    "/api/v1/places/book_requests": {
      "get": {
        "tags": ["Booking Requests"], "summary": "List incoming booking requests for the logged-in owner's/guide's places",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": { "description": "Booking requests with place_name/traveller_name added", "content": { "application/json": { "schema": {"type": "array", "items": {"$ref": "#/components/schemas/BookingRequest"}}}}}
        }
      }
    },
...
  }
}